fix(mobile): fix the recipient address validation helper
